Fix parsing for power
When testing two variable expressions, found that multiplication after
power e.g. "x^2*y^3" was not working. This was due to an error in the
parsing algorithm which was emptying the operator stack completely when
it shouldn't have been. Fixing that then resolved the error in the
evaluation. Also added the case as a test for future.